requirements.txt
requests==2.26.0
ply==3.11
python-dotenv==1.0.1
google-generativeai==0.1.0rc1